Add files via upload
What I have been working on. I had tried a few methods, but there was a lot of errors due to the size of the data so I split the data to look at just the lung cancer cells by joining the labels with the data and filtering. I'll be working on retrying the methods and visualizing the data afterwards and before as one means of comparison.