Updated, organized, and new section I'll work on.
I have all the data sets that are worked on within the file if any of you need them. Additionally, I added the feature selection section in order to compare to the auto encoder which is what we planned to do with all the other steps. The prof.  mentioned we'll need to know the features in order to determine the biological implications so it's also necessary to work on so we can include it in our report.