Updated and organized. Working on feat. selection
Finished denoising and Magic seems to work best. As I mentioned in my last commit, I think it'd be useful to do feat. selection and compare it with the features that the auto-encoder chooses. I already got 10 features that are supposedly important and I wrote down what they are and their biological implications. Will continue on this work.